feat: read drone status

+++ src/main/kotlin/drone/DroneStatus.kt
@@ -3,4 +3,4 @@ package drone
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class DroneStatus(val state: DroneState, val battery: Int?, val location: String?)
+data class DroneStatus(val state: DroneState, val battery: Float?, val location: String?)
